+++ packages/remix-cloudflare-pages/CHANGELOG.md
@@ -1,5 +1,54 @@
 # `@remix-run/cloudflare-pages`
 
+## 1.7.5
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.7.5`
+
+## 1.7.4
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.7.4`
+
+## 1.7.3
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.7.3`
+
+## 1.7.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.7.2`
+
+## 1.7.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.7.1`
+
+## 1.7.0
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.7.0`
+
+## 1.6.8
+
+### Patch Changes
+
+- Updated dependencies:
+  - `@remix-run/cloudflare@1.6.8`
+
 ## 1.6.7
 
 ### Patch Changes